+++ src/infra/database/database.module.ts
@@ -2,6 +2,7 @@ import { Global, Module } from '@nestjs/common';
 
 import { PrismaService } from './prisma/prisma.service';
 import { CityRepository } from './prisma/repositories/city/city.repository';
+import { CommentRepository } from './prisma/repositories/comment/comment.repository';
 import { FeatureFlagRepository } from './prisma/repositories/feature-flag/feature-flag.repository';
 import { IssueRepository } from './prisma/repositories/issue/issue.repository';
 import { UserRepository } from './prisma/repositories/user/user.repository';
@@ -14,6 +15,7 @@ import { UserRepository } from './prisma/repositories/user/user.repository';
     FeatureFlagRepository,
     CityRepository,
     IssueRepository,
+    CommentRepository,
   ],
   exports: [
     PrismaService,
@@ -21,6 +23,7 @@ import { UserRepository } from './prisma/repositories/user/user.repository';
     FeatureFlagRepository,
     CityRepository,
     IssueRepository,
+    CommentRepository,
   ],
 })
 export class DatabaseModule {}