+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [FMC-4](https://fixed-my-city.atlassian.net/browse/FMC-4) - Criar módulo de issue
 - [FMC-5](https://fixed-my-city.atlassian.net/browse/FMC-5) - Criar funcionalidade de listar feature flags
 - [FMC-6](https://fixed-my-city.atlassian.net/browse/FMC-6) - Expirar token de resetar senha
+- [FMC-8](https://fixed-my-city.atlassian.net/browse/FMC-8) - Criar Módulo de comment
 - [FMC-9](https://fixed-my-city.atlassian.net/browse/FMC-9) - Ajustar tudo que envolve cidade para receber cityId como parametro quando é ADMIN
 
 ### Changed